refactor: :recycle: Refactor agreement review workflow and add Services Component
